try pure ts project
