--- Marklet.js
@@ -1,5 +1,0 @@
-const Lexer = require('./Lexer')
-
-new Lexer()
-
-console.log()
