+++ src/Marklet.ts
@@ -1,0 +1,3 @@
+import Document from './Document'
+
+console.log(new Document().parse('aaa'))
