+++ tsconfig.json
@@ -4,9 +4,10 @@
         "target": "es6",
         "noImplicitAny": false,
         "sourceMap": true,
-        "allowJs": true
+        "allowJs": true,
+        "outDir": "dist"
     },
-    "exclude": [
-        "node_modules"
+    "include": [
+        "src"
     ]
 }
